+++ packages/autoconfigure/src/index.ts
@@ -139,7 +139,7 @@ import {
   resolveTasksFile
 } from "./personal-providers.js";
 
-export { buildVoiceRegistry } from "./personal-providers.js";
+export { buildVoiceRegistry, resolveNotesDir, resolveTasksFile } from "./personal-providers.js";
 import {
   createConversationSummaryStore,
   createDebugReplayCaptureStore,